Extracted the 'GetMaxMemoryAndTimeUsed' into a helper class.
